+++ cpp/include/graph.hpp
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 #pragma once
-#include <comms_mpi.hpp>
-#include <iostream>
-#include <memory>
-#include <rmm/device_buffer.hpp>
-
 #include <cstddef>
 #include <cstdint>
+#include <iostream>
+#include <memory>
+#include <raft/handle.hpp>
+#include <rmm/device_buffer.hpp>
 
 namespace cugraph {
 namespace experimental {
@@ -55,8 +54,7 @@ template <typename VT, typename ET, typename WT>
 class GraphViewBase {
  public:
   WT *edge_data;  ///< edge weight
-  Comm comm;
-
+  raft::handle_t *handle;
   GraphProperties prop;
 
   VT number_of_vertices;
@@ -69,15 +67,14 @@ class GraphViewBase {
    * identifiers
    */
   void get_vertex_identifiers(VT *identifiers) const;
-  void set_communicator(Comm &comm_) { comm = comm_; }
-
+  void set_handle(raft::handle_t *handle_) { handle = handle_; }
   GraphViewBase(WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
     : edge_data(edge_data_),
-      comm(),
       prop(),
       number_of_vertices(number_of_vertices_),
       number_of_edges(number_of_edges_)
   {
+    handle = new raft::handle_t;
   }
   bool has_data(void) const { return edge_data != nullptr; }
 };
@@ -100,7 +97,8 @@ class GraphCOOView : public GraphViewBase<VT, ET, WT> {
    *
    * @throws     cugraph::logic_error when an error occurs.
    *
-   * @param[out] degree                Device array of size V (V is number of vertices) initialized
+   * @param[out] degree                Device array of size V (V is number of
+   * vertices) initialized
    * to zeros. Will contain the computed degree of every vertex.
    * @param[in]  direction             IN_PLUS_OUT, IN or OUT
    */
@@ -114,15 +112,20 @@ class GraphCOOView : public GraphViewBase<VT, ET, WT> {
   /**
    * @brief      Wrap existing arrays representing an edge list in a Graph.
    *
-   *             GraphCOOView does not own the memory used to represent this graph. This
+   *             GraphCOOView does not own the memory used to represent this
+   * graph. This
    *             function does not allocate memory.
    *
-   * @param  source_indices        This array of size E (number of edges) contains the index of the
+   * @param  source_indices        This array of size E (number of edges)
+   * contains the index of the
    * source for each edge. Indices must be in the range [0, V-1].
-   * @param  destination_indices   This array of size E (number of edges) contains the index of the
+   * @param  destination_indices   This array of size E (number of edges)
+   * contains the index of the
    * destination for each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array size E (number of edges) contains the weight for each
-   * edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array size E (number of edges) contains
+   * the weight for each
+   * edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -136,7 +139,8 @@ class GraphCOOView : public GraphViewBase<VT, ET, WT> {
 };
 
 /**
- * @brief       Base class for graph stored in CSR (Compressed Sparse Row) format or CSC (Compressed
+ * @brief       Base class for graph stored in CSR (Compressed Sparse Row)
+ * format or CSC (Compressed
  * Sparse Column) format
  *
  * @tparam VT   Type of vertex id
@@ -163,9 +167,11 @@ class GraphCompressedSparseBaseView : public GraphViewBase<VT, ET, WT> {
    *
    * @throws     cugraph::logic_error when an error occurs.
    *
-   * @param[out] degree                Device array of size V (V is number of vertices) initialized
+   * @param[out] degree                Device array of size V (V is number of
+   * vertices) initialized
    * to zeros. Will contain the computed degree of every vertex.
-   * @param[in]  x                     Integer value indicating type of degree calculation
+   * @param[in]  x                     Integer value indicating type of degree
+   * calculation
    *                                      0 : in+out degree
    *                                      1 : in-degree
    *                                      2 : out-degree
@@ -174,16 +180,22 @@ class GraphCompressedSparseBaseView : public GraphViewBase<VT, ET, WT> {
 
   /**
    * @brief      Wrap existing arrays representing adjacency lists in a Graph.
-   *             GraphCSRView does not own the memory used to represent this graph. This
+   *             GraphCSRView does not own the memory used to represent this
+   * graph. This
    *             function does not allocate memory.
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -213,16 +225,22 @@ class GraphCSRView : public GraphCompressedSparseBaseView<VT, ET, WT> {
 
   /**
    * @brief      Wrap existing arrays representing adjacency lists in a Graph.
-   *             GraphCSRView does not own the memory used to represent this graph. This
+   *             GraphCSRView does not own the memory used to represent this
+   * graph. This
    *             function does not allocate memory.
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -250,17 +268,24 @@ class GraphCSCView : public GraphCompressedSparseBaseView<VT, ET, WT> {
   GraphCSCView() : GraphCompressedSparseBaseView<VT, ET, WT>(nullptr, nullptr, nullptr, 0, 0) {}
 
   /**
-   * @brief      Wrap existing arrays representing transposed adjacency lists in a Graph.
-   *             GraphCSCView does not own the memory used to represent this graph. This
+   * @brief      Wrap existing arrays representing transposed adjacency lists in
+   * a Graph.
+   *             GraphCSCView does not own the memory used to represent this
+   * graph. This
    *             function does not allocate memory.
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -273,14 +298,19 @@ class GraphCSCView : public GraphCompressedSparseBaseView<VT, ET, WT> {
 };
 
 /**
- * @brief      TODO : Change this Take ownership of the provided graph arrays in COO format
+ * @brief      TODO : Change this Take ownership of the provided graph arrays in
+ * COO format
  *
- * @param  source_indices        This array of size E (number of edges) contains the index of the
+ * @param  source_indices        This array of size E (number of edges) contains
+ * the index of the
  * source for each edge. Indices must be in the range [0, V-1].
- * @param  destination_indices   This array of size E (number of edges) contains the index of the
+ * @param  destination_indices   This array of size E (number of edges) contains
+ * the index of the
  * destination for each edge. Indices must be in the range [0, V-1].
- * @param  edge_data             This array size E (number of edges) contains the weight for each
- * edge.  This array can be null in which case the graph is considered unweighted.
+ * @param  edge_data             This array size E (number of edges) contains
+ * the weight for each
+ * edge.  This array can be null in which case the graph is considered
+ * unweighted.
  * @param  number_of_vertices    The number of vertices in the graph
  * @param  number_of_edges       The number of edges in the graph
  */
@@ -314,12 +344,16 @@ class GraphCOO {
   /**
    * @brief      Take ownership of the provided graph arrays in COO format
    *
-   * @param  source_indices        This array of size E (number of edges) contains the index of the
+   * @param  source_indices        This array of size E (number of edges)
+   * contains the index of the
    * source for each edge. Indices must be in the range [0, V-1].
-   * @param  destination_indices   This array of size E (number of edges) contains the index of the
+   * @param  destination_indices   This array of size E (number of edges)
+   * contains the index of the
    * destination for each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array size E (number of edges) contains the weight for each
-   * edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array size E (number of edges) contains
+   * the weight for each
+   * edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -389,7 +423,8 @@ struct GraphSparseContents {
 };
 
 /**
- * @brief       Base class for constructted graphs stored in CSR (Compressed Sparse Row) format or
+ * @brief       Base class for constructted graphs stored in CSR (Compressed
+ * Sparse Row) format or
  * CSC (Compressed Sparse Column) format
  *
  * @tparam VT   Type of vertex id
@@ -410,13 +445,18 @@ class GraphCompressedSparseBase {
   /**
    * @brief      Take ownership of the provided graph arrays in CSR/CSC format
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -466,7 +506,8 @@ class GraphCompressedSparseBase {
 };
 
 /**
- * @brief       A constructed graph stored in CSR (Compressed Sparse Row) format.
+ * @brief       A constructed graph stored in CSR (Compressed Sparse Row)
+ * format.
  *
  * @tparam VT   Type of vertex id
  * @tparam ET   Type of edge id
@@ -483,13 +524,18 @@ class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
   /**
    * @brief      Take ownership of the provided graph arrays in CSR format
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
@@ -519,7 +565,8 @@ class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
 };
 
 /**
- * @brief       A constructed graph stored in CSC (Compressed Sparse Column) format.
+ * @brief       A constructed graph stored in CSC (Compressed Sparse Column)
+ * format.
  *
  * @tparam VT   Type of vertex id
  * @tparam ET   Type of edge id
@@ -536,13 +583,18 @@ class GraphCSC : public GraphCompressedSparseBase<VT, ET, WT> {
   /**
    * @brief      Take ownership of the provided graph arrays in CSR format
    *
-   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
-   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * @param  offsets               This array of size V+1 (V is number of
+   * vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0,
+   * E] (number of
    * edges).
-   * @param  indices               This array of size E contains the index of the destination for
+   * @param  indices               This array of size E contains the index of
+   * the destination for
    * each edge. Indices must be in the range [0, V-1].
-   * @param  edge_data             This array of size E (number of edges) contains the weight for
-   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  edge_data             This array of size E (number of edges)
+   * contains the weight for
+   * each edge.  This array can be null in which case the graph is considered
+   * unweighted.
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */