+++ python/cugraph/traversal/sssp_wrapper.pyx
@@ -148,5 +148,5 @@ def sssp(input_graph, source):
             df = unrenumbered_df[cols[1:n_cols + 1] + [cols[0]] + cols[n_cols:]]
         else: # Simple renumbering
             df = unrenumber(input_graph.edgelist.renumber_map, df, 'vertex')
-            df['predecessor'][df['predecessor'] >- 1] = input_graph.edgelist.renumber_map[df['predecessor'][df['predecessor'] >- 1]]
+            df['predecessor'][df['predecessor'] >- 1] = input_graph.edgelist.renumber_map.iloc[df['predecessor'][df['predecessor'] >- 1]]
     return df